agent/requirements.txt
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-community==0.3.7
anthropic==0.40.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0
